services/video-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
numpy==1.26.4
vaderSentiment==3.3.2
matplotlib==3.9.2
Pillow==10.4.0
pyttsx3==2.98
transformers==4.44.2
imageio-ffmpeg==0.5.1
requests==2.32.3
beautifulsoup4==4.12.3
sentence-transformers==3.0.1
langchain==0.3.7
langchain-community==0.3.7
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langchain-chroma==0.1.4
chromadb==0.5.15
diffusers==0.30.3
accelerate==0.34.2
moviepy==1.0.3
gTTS==2.5.4
